#1499 Defining two Standard Request Type associated to the same table, event model validator and different where clause, the requests defined in each one duplicates if the conditions are ok.  https://github.com/adempiere/adempiere/issues/1499 - Refactor the methods move to model class and implement other approach to add listener model validator
